third_party/WebKit/LayoutTests/fast/events/mouseevent-getModifierState-expected.txt

Tests MouseEvent.getModifierState()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS lastMouseEvent.type is "mousedown"
PASS lastMouseEvent.button is 0
PASS typeof lastMouseEvent.getModifierState('Control') is 'boolean'
PASS lastMouseEvent.getModifierState('Control') is true
PASS typeof lastMouseEvent.ctrlKey is 'boolean'
PASS lastMouseEvent.ctrlKey is true
PASS typeof lastMouseEvent.getModifierState('Shift') is 'boolean'
PASS lastMouseEvent.getModifierState('Shift') is false
PASS typeof lastMouseEvent.shiftKey is 'boolean'
PASS lastMouseEvent.shiftKey is false
PASS typeof lastMouseEvent.getModifierState('Meta') is 'boolean'
PASS lastMouseEvent.getModifierState('Meta') is false
PASS typeof lastMouseEvent.metaKey is 'boolean'
PASS lastMouseEvent.metaKey is false
PASS typeof lastMouseEvent.getModifierState('Alt') is 'boolean'
PASS lastMouseEvent.getModifierState('Alt') is false
PASS typeof lastMouseEvent.altKey is 'boolean'
PASS lastMouseEvent.altKey is false
PASS lastMouseEvent.type is "mousedown"
PASS lastMouseEvent.button is 1
PASS typeof lastMouseEvent.getModifierState('Control') is 'boolean'
PASS lastMouseEvent.getModifierState('Control') is false
PASS typeof lastMouseEvent.ctrlKey is 'boolean'
PASS lastMouseEvent.ctrlKey is false
PASS typeof lastMouseEvent.getModifierState('Shift') is 'boolean'
PASS lastMouseEvent.getModifierState('Shift') is true
PASS typeof lastMouseEvent.shiftKey is 'boolean'
PASS lastMouseEvent.shiftKey is true
PASS typeof lastMouseEvent.getModifierState('Meta') is 'boolean'
PASS lastMouseEvent.getModifierState('Meta') is false
PASS typeof lastMouseEvent.metaKey is 'boolean'
PASS lastMouseEvent.metaKey is false
PASS typeof lastMouseEvent.getModifierState('Alt') is 'boolean'
PASS lastMouseEvent.getModifierState('Alt') is false
PASS typeof lastMouseEvent.altKey is 'boolean'
PASS lastMouseEvent.altKey is false
PASS lastMouseEvent.type is "mousedown"
PASS lastMouseEvent.button is 2
PASS typeof lastMouseEvent.getModifierState('Control') is 'boolean'
PASS lastMouseEvent.getModifierState('Control') is false
PASS typeof lastMouseEvent.ctrlKey is 'boolean'
PASS lastMouseEvent.ctrlKey is false
PASS typeof lastMouseEvent.getModifierState('Shift') is 'boolean'
PASS lastMouseEvent.getModifierState('Shift') is false
PASS typeof lastMouseEvent.shiftKey is 'boolean'
PASS lastMouseEvent.shiftKey is false
PASS typeof lastMouseEvent.getModifierState('Meta') is 'boolean'
PASS lastMouseEvent.getModifierState('Meta') is false
PASS typeof lastMouseEvent.metaKey is 'boolean'
PASS lastMouseEvent.metaKey is false
PASS typeof lastMouseEvent.getModifierState('Alt') is 'boolean'
PASS lastMouseEvent.getModifierState('Alt') is true
PASS typeof lastMouseEvent.altKey is 'boolean'
PASS lastMouseEvent.altKey is true
PASS lastMouseEvent.type is "mousedown"
PASS lastMouseEvent.button is 0
PASS typeof lastMouseEvent.getModifierState('Control') is 'boolean'
PASS lastMouseEvent.getModifierState('Control') is false
PASS typeof lastMouseEvent.ctrlKey is 'boolean'
PASS lastMouseEvent.ctrlKey is false
PASS typeof lastMouseEvent.getModifierState('Shift') is 'boolean'
PASS lastMouseEvent.getModifierState('Shift') is false
PASS typeof lastMouseEvent.shiftKey is 'boolean'
PASS lastMouseEvent.shiftKey is false
PASS typeof lastMouseEvent.getModifierState('Meta') is 'boolean'
PASS lastMouseEvent.getModifierState('Meta') is true
PASS typeof lastMouseEvent.metaKey is 'boolean'
PASS lastMouseEvent.metaKey is true
PASS typeof lastMouseEvent.getModifierState('Alt') is 'boolean'
PASS lastMouseEvent.getModifierState('Alt') is false
PASS typeof lastMouseEvent.altKey is 'boolean'
PASS lastMouseEvent.altKey is false
PASS lastMouseEvent.type is "mousedown"
PASS lastMouseEvent.button is 1
PASS typeof lastMouseEvent.getModifierState('Control') is 'boolean'
PASS lastMouseEvent.getModifierState('Control') is true
PASS typeof lastMouseEvent.ctrlKey is 'boolean'
PASS lastMouseEvent.ctrlKey is true
PASS typeof lastMouseEvent.getModifierState('Shift') is 'boolean'
PASS lastMouseEvent.getModifierState('Shift') is true
PASS typeof lastMouseEvent.shiftKey is 'boolean'
PASS lastMouseEvent.shiftKey is true
PASS typeof lastMouseEvent.getModifierState('Meta') is 'boolean'
PASS lastMouseEvent.getModifierState('Meta') is false
PASS typeof lastMouseEvent.metaKey is 'boolean'
PASS lastMouseEvent.metaKey is false
PASS typeof lastMouseEvent.getModifierState('Alt') is 'boolean'
PASS lastMouseEvent.getModifierState('Alt') is false
PASS typeof lastMouseEvent.altKey is 'boolean'
PASS lastMouseEvent.altKey is false
PASS lastMouseEvent.type is "mousedown"
PASS lastMouseEvent.button is 2
PASS typeof lastMouseEvent.getModifierState('Control') is 'boolean'
PASS lastMouseEvent.getModifierState('Control') is true
PASS typeof lastMouseEvent.ctrlKey is 'boolean'
PASS lastMouseEvent.ctrlKey is true
PASS typeof lastMouseEvent.getModifierState('Shift') is 'boolean'
PASS lastMouseEvent.getModifierState('Shift') is true
PASS typeof lastMouseEvent.shiftKey is 'boolean'
PASS lastMouseEvent.shiftKey is true
PASS typeof lastMouseEvent.getModifierState('Meta') is 'boolean'
PASS lastMouseEvent.getModifierState('Meta') is false
PASS typeof lastMouseEvent.metaKey is 'boolean'
PASS lastMouseEvent.metaKey is false
PASS typeof lastMouseEvent.getModifierState('Alt') is 'boolean'
PASS lastMouseEvent.getModifierState('Alt') is true
PASS typeof lastMouseEvent.altKey is 'boolean'
PASS lastMouseEvent.altKey is true
PASS lastMouseEvent.type is "mousedown"
PASS lastMouseEvent.button is 1
PASS typeof lastMouseEvent.getModifierState('Control') is 'boolean'
PASS lastMouseEvent.getModifierState('Control') is true
PASS typeof lastMouseEvent.ctrlKey is 'boolean'
PASS lastMouseEvent.ctrlKey is true
PASS typeof lastMouseEvent.getModifierState('Shift') is 'boolean'
PASS lastMouseEvent.getModifierState('Shift') is true
PASS typeof lastMouseEvent.shiftKey is 'boolean'
PASS lastMouseEvent.shiftKey is true
PASS typeof lastMouseEvent.getModifierState('Meta') is 'boolean'
PASS lastMouseEvent.getModifierState('Meta') is true
PASS typeof lastMouseEvent.metaKey is 'boolean'
PASS lastMouseEvent.metaKey is true
PASS typeof lastMouseEvent.getModifierState('Alt') is 'boolean'
PASS lastMouseEvent.getModifierState('Alt') is true
PASS typeof lastMouseEvent.altKey is 'boolean'
PASS lastMouseEvent.altKey is true
PASS successfullyParsed is true

TEST COMPLETE
